pyproject.toml
[project]
name = "docling-langextract-testing"
version = "0.2.0"
description = "Testing docling + langextract combination for paralegal date extraction"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40.0",
    "docling>=0.1.0",
    "duckdb>=1.0.0",
    "extract-msg>=0.55.0",
    "fastapi>=0.115.0",
"flask>=3.0.0",
"flask-session>=0.8.0",
"flask-wtf>=1.2.0",
"werkzeug>=3.0.0",
    "google-generativeai>=0.8.0",
    "langextract",
    "numpy<2",
    "ocrmac>=1.0.0",
    "openai>=1.0.0",
    "openpyxl>=3.1.5",
    "paddleocr>=3.2.0",
    "pandas>=2.3.2",
    "pdf2image>=1.17.0",
    "plotly>=6.3.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "pymupdf>=1.24.0",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.1",
    "rapidocr-onnxruntime>=1.4.4",
    "regex>=2024.9.11",
    "requests>=2.32.5",
    "streamlit>=1.49.1",
    "tesserocr>=2.8.0",
    "textstat>=0.7.10",
    "tiktoken>=0.7.0",
    "torch>=2.0.0,<2.3.0",
    "uvicorn[standard]>=0.32.0",
]

[tool.uv.sources]
langextract = { git = "https://github.com/google/langextract.git" }
